Rename clien4.d.ts to client.d.ts
